Upgrade dependencies
Motivation
------------
Several dependencies are out of date.

Modifications
---------------
Upgraded to Node 12.

Upgraded to Couchbase SDK 2.6.8 and couchbase-index-manager 0.7.2.

Hacked package-lock.json to make fakeit use the latest Couchbase SDK. couchbase-promises has a poor dependency on "~2.4.0" which was causing a Couchbase SDK version fallback, and the old version wouldn't compile successfully on the Docker image.